+++ test/unit/render_cache.test.js
@@ -24,6 +24,8 @@ suite('render_cache', function() {
       });
     });
 
+if (0) { // DROPME
+
     test('has a cache of render objects', function(){
         var render_cache = new RenderCache(10000, mml_store);
         assert.ok(_.isObject(render_cache.renderers));
@@ -71,6 +73,8 @@ suite('render_cache', function() {
         });
     });
 
+} // DROPME
+
     test('cache renderer hook is only called when a _new_ cache is created', function(done){
         var render_cache = new RenderCache(10000, mml_store);
         var req = {params: {dbname: "windshaft_test", table: 'test_table', x: 4, y:4, z:4, geom_type:'point', style:"#test_table{}", format:'png', style_version:'2.1.0' }};
@@ -102,6 +106,8 @@ suite('render_cache', function() {
           }
         );
     });
+
+if (0) { // DROPME
 
     test('cache renderer item contains cache_buster', function(done){
         var render_cache = new RenderCache(10000, mml_store);
@@ -269,6 +275,8 @@ suite('render_cache', function() {
             setTimeout(function(){assert.equal(_.keys(render_cache.renderers).length, 0); done();},200);
         });
     });
+
+} // DROPME
 
     suiteTeardown(function(done) {
       // Flush redis cache